Move asset-mapping.js out of cumulus dir
